requirements.txt
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-Login==0.5.0
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
peewee==3.14.8
psycopg2==2.9.2
psycopg2-binary==2.9.2
pycparser==2.21
python-dotenv==0.19.2
requests==2.26.0
six==1.16.0
urllib3==1.26.7
Werkzeug==2.0.2
